[#103] refactor: 폴더 인식으로 인한 modal 폴더명 수정
