+++ src/app/_components/preview/menu/RaidSelecterModal.tsx
@@ -2,12 +2,12 @@
 import { useState } from "react";
 import Image from "next/image";
 
-import Modal from "../../common/modal/Modal";
 import { Button } from "@headlessui/react";
 import { useRaidStore } from "@/app/_store/raidStore";
 import BottomArrow from "../../common/BottomArrow";
 import LimitRaidSelecter from "../table/LimitRaidSelecter";
 import RaidSelecter from "../table/RaidSelecter";
+import Modal from "../../common/modalR/Modal";
 
 const RaidSelecterModal = ({ isAllBtn = true }: { isAllBtn: boolean }) => {
   const [open, setOpen] = useState(false);
